small fixes to category
