+++ frontend/src/App.jsx
@@ -10,14 +10,12 @@ import './App.css';
 import IndexRouter from './route/IndexRouter';
 import { apiURL } from './config/env';
 import authenticationReducer from './store/reducers/auth';
-import jesseReducer from './store/reducers/jesse';
 
 export const history = createBrowserHistory({ basename: '/' });
 
 
 const reducer = (history) => combineReducers({
   authenticationReducer,
-  jesseReducer,
   router: connectRouter(history),
 });
 
@@ -31,7 +29,7 @@ function App() {
       <Provider store={store}>
         <ConnectedRouter history={history}>
           <BrowserRouter>
-            <IndexRouter/>
+            <IndexRouter />
           </BrowserRouter>
         </ConnectedRouter>
       </Provider>
