+++ frontend/src/pages/Dashboard.jsx
@@ -1,16 +1,12 @@
 import React, { useState } from 'react';
-import Header from '../components/layout/Header'
-import { useSelector } from 'react-redux';
-import UserProfile from './UserProfile';
+import Header from '../components/layout/Header';
+import UserCategory from '../components/UserCategory';
 
 function Dashboard() {
-  const user = useSelector(state => state.jesseReducer.user.user);
-  const [errorMessage, setErrorMessage] = useState('');
-
   return (
     <div>
       <Header></Header>
-      <UserProfile />
+      <UserCategory />
     </div>
   );
 }
