--- frontend/src/store/action/jesse.js
@@ -1,6 +1,0 @@
-export const jesseRedux = (user) => {
-    return {
-        type: 'JESSE',
-        user
-    }
-}
